pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "omniduct"
dynamic = ["version"]
description = "A toolkit providing a uniform interface for connecting to and extracting data from a wide variety of (potentially remote) data stores (including HDFS, Hive, Presto, MySQL, etc)."
readme = "README.md"
license = "MIT"
authors = [
    { name = "Matthew Wardrop", email = "mpwardrop@gmail.com" },
    { name = "Dan Frank", email = "danfrankj@gmail.com" },
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
requires-python = ">=3.10"
dependencies = [
    "decorator",
    "interface_meta>=1.2.0,<2",
    "jinja2",
    "lazy-object-proxy",
    "packaging",
    "typing_extensions>=4.0.0",
    "pandas>=0.20.3",
    "progressbar2>=3.30.0",
    "python-dateutil",
    "pyyaml",
    "sqlalchemy",
    "sqlparse",
    "wrapt",
]

[project.optional-dependencies]
all = [
    "boto3",
    "coverage",
    "flake8",
    "mock",
    "nose",
    "paramiko",
    "pexpect",
    "pexpect",
    "pydruid>=0.4.0",
    "pyexasol",
    "pyfakefs",
    "pyhive[hive]>=0.4",
    "pyhive[presto]>=0.4",
    "pyspark",
    "pywebhdfs",
    "requests",
    "requests",
    "snowflake-sqlalchemy",
    "sphinx",
    "sphinx_autobuild",
    "sphinx_rtd_theme",
    "thrift>=0.10.0",
]
docs = [
    "sphinx",
    "sphinx_autobuild",
    "sphinx_rtd_theme",
]
druid = [
    "pydruid>=0.4.0",
]
exasol = [
    "pyexasol",
]
hiveserver2 = [
    "pyhive[hive]>=0.4",
    "thrift>=0.10.0",
]
presto = [
    "pyhive[presto]>=0.4",
]
pyspark = [
    "pyspark",
]
rest = [
    "requests",
]
s3 = [
    "boto3",
]
snowflake = [
    "snowflake-sqlalchemy",
]
ssh = [
    "pexpect",
]
ssh_paramiko = [
    "paramiko",
    "pexpect",
]
test = [
    "coverage",
    "flake8",
    "mock",
    "nose",
    "pyfakefs",
]
webhdfs = [
    "pywebhdfs",
    "requests",
]

[project.urls]
Homepage = "https://github.com/airbnb/omniduct"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "omniduct/_version_info.py"

# Build target configuration

[tool.hatch.build.targets.sdist]
include = [
    "docs",
    "example_wrapper",
    "omniduct",
    "tests",
    "LICENSE",
    "README.md",
    "MANIFEST.in",
    "pyproject.toml",
]

# Testing configuration

[tool.hatch.envs.default]
dependencies = [
    "mock",
    "pyfakefs",
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "requests",
]

[tool.hatch.envs.default.scripts]
tests = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=omniduct --cov-report=xml -vv {args:tests}"

[[tool.hatch.envs.test.matrix]]
python = ["310", "311", "312", "313", "314"]

[tool.hatch.envs.lint]
detached=true
dependencies = [
    "ruff==0.15.6",
    # Type-checking deps
    "mypy==1.19.1",
    "pandas-stubs",
    "types-python-dateutil",
    "types-PyYAML",
    "types-requests",
    "typing_extensions",
]

[tool.hatch.envs.lint.scripts]
check = [
    "ruff check omniduct tests",
    "ruff format --check omniduct tests",
    "mypy omniduct",
]
format = [
    "ruff check --fix omniduct tests",
    "ruff format omniduct tests",
]

# Linter and format configuration


[tool.ruff.lint]
select = [
    "F",  # flake8 rules
    "E",  # pycodestyle
    "W",  # pycodestyle
    "S",  # bandit
    "I001",  # import sorting
    # "D", # documentation
    "UP",  # migrate code to newer conventions when we bump Python versions
]
ignore = [
    "E501",  # Allow long lines
    "E722",  # Allow bare 'except'
    "S110",  # Allow try-except-pass for connection cleanup
    "S603",  # Allow use of subprocess module
    "S608",  # Allow string-based SQL construction in database clients
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "S101",  # Allow use of assert in tests
]

# Mypy type-checking configuration

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true

# Optional dependencies that ship no type stubs
[[tool.mypy.overrides]]
module = [
    "boto3",
    "botocore",
    "botocore.*",
    "decorator",
    "jinja2",
    "jinja2.*",
    "impala",
    "impala.*",
    "interface_meta",
    "interface_meta.*",
    "IPython",
    "IPython.*",
    "lazy_object_proxy",
    "neo4j.*",
    "opinel.*",
    "progressbar2",
    "progressbar",
    "pyarrow.*",
    "TCLIService.*",
    "wrapt",
    "sqlparse",
    "pexpect",
    "pydruid.*",
    "pyhive.*",
    "pywebhdfs.*",
    "pyexasol",
    "pyspark.*",
    "thrift.*",
    "packaging",
    "packaging.*",
    "paramiko",
    "snowflake.*",
    "sqlalchemy",
    "sqlalchemy.*",
]
ignore_missing_imports = true
